+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2025 Aliano Manuel, Ávila Simón, Iturria Agustín, Nicolás Seijo, Victor Maza
+Copyright (c) 2025 Aliano Manuel, Ávila Simón, Iturria Agustín, Seijo Nicolás, Maza Victor
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
