+++ main.py
@@ -1,0 +1,18 @@
+# ==============================================================================
+# Materia: Programación 1 
+# Grupo 6: Aliano Manuel, Ávila Simón, Iturria Agustín, Seijo Nicolás, Maza Victor
+# Descripción: Sistema CRUD de Reserva de turnos médicos
+# Fecha de creación: 10/08/2025
+# ==============================================================================
+
+import pacientes
+import turnos
+import medicos
+import aux
+
+# ==============================================================================
+# ==============================PROGRAMA PRINCIPAL==============================
+# ==============================================================================
+
+aux.grupo6_dev_logo()
+
