+++ pacientes.py
@@ -1,0 +1,9 @@
+# ==============================================================================
+# Materia: Programación 1 
+# Grupo 6: Aliano Manuel, Ávila Simón, Iturria Agustín, Nicolás Seijo, Victor Maza
+# Descripción: 
+# Fecha de creación: 10/08/2025
+# ==============================================================================
+
+print("l")
+
