+++ turnos.py
@@ -1,0 +1,11 @@
+# ==============================================================================
+# Materia: Programación 1 
+# Grupo 6: Aliano Manuel, Ávila Simón, Iturria Agustín, Nicolás Seijo, Victor Maza
+# Descripción: 
+# Fecha de creación: 10/08/2025
+# ==============================================================================
+
+ENCABEZADO = []# Lista de títulos de la matriz de datos. Ejemplo en el googlesheet. 
+ESTADO_TURNO = []# Lista de títulos de la matriz de datos. Ejemplo en el googlesheet.
+ESPECIALIDAD = []# Lista de títulos de la matriz de datos. Ejemplo en el googlesheet.
+
